apps/hash-agents/pyproject.toml
[tool.poetry]
name = "hash-agents"
version = "0.0.0"
description = ""
license = "AGPL-3.0-only"
authors = ["HASH"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.137"
openai = "^0.27.4"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
tiktoken = "^0.3.3"
beartype = "^0.13.1"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"


[tool.poetry.group.dev.dependencies]
setuptools = "^67.7.1"

[tool.poetry.group.tools.dependencies]
black = "^23.3.0"
ruff = "^0.0.262"
isort = {extras = ["colors"], version = "^5.12.0"}

[tool.poetry.group.production.dependencies]
gunicorn = "^20.1.0"


[tool.black]
target-version = ['py311']
preview = true

[tool.ruff]
select = ["ANN", "B", "C901", "E", "F", "G201", "N", "UP", "S", "TRY"]
ignore = ["ANN101"]

[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
